+++ src/js/util/lang.js
@@ -227,10 +227,10 @@ export function intersectRect(r1, r2) {
 }
 
 export function pointInRect(point, rect) {
-    return point.x <= r2.right &&
-        point.x >= r2.left &&
-        point.y <= r2.bottom &&
-        point.y >= r2.top;
+    return point.x <= rect.right &&
+        point.x >= rect.left &&
+        point.y <= rect.bottom &&
+        point.y >= rect.top;
 }
 
 export const Dimensions = {